+++ profapp/__init__.py
@@ -1,17 +1,18 @@
-#from .db import db
 from flask import Flask
-from profapp.controllers.views import index
 from profapp.controllers.blueprints import user_bp, article_bp, filemanager_bp
+import profapp.controllers.views_article as views_art
+import profapp.controllers.views_auth as views_auth
+import profapp.controllers.views_filemanager as views_fileman
+import profapp.controllers.views_index as views_ind
 
 def create_app(config='config.ProductionDevelopmentConfig'):
     app = Flask(__name__)
     app.config.from_object(config)
-    #db.init_app(app)
 
-    app.add_url_rule('/', 'index', index)
-
-    app.register_blueprint(article_bp, url_prefix='/articles')
-    app.register_blueprint(user_bp, url_prefix='/users')
-    app.register_blueprint(filemanager_bp, url_prefix='/filemanager')
+    app.register_blueprint(views_ind.general_bp, url_prefix='/')
+    app.register_blueprint(views_art.article_bp, url_prefix='/articles')
+    app.register_blueprint(views_auth.user_bp, url_prefix='/users')
+    app.register_blueprint(views_fileman.filemanager_bp,
+                           url_prefix='/filemanager')
 
     return app
