+++ profapp/controllers/blueprints.py
@@ -1,5 +1,6 @@
 from flask import Blueprint
 
+general_bp = Blueprint('general', __name__)
 user_bp = Blueprint('user', __name__)
 article_bp = Blueprint('articles', __name__)
 filemanager_bp = Blueprint('filemanager', __name__)
