+++ profapp/controllers/views_index.py
@@ -1,0 +1,6 @@
+from flask import render_template
+from .blueprints import general_bp
+
+@general_bp.route('/')
+def index():
+    return render_template('index.html')
