+++ profapp/templates/filemanager.html
@@ -17,7 +17,7 @@
   <script type="text/javascript" src="/static/filemanager/bower_components/angular-filemanager/dist/angular-filemanager.min.js"></script>
   <script type="text/javascript" src="/static/filemanager/bower_components/angular-filemanager/dist/cached-templates.js"></script>
   <link type="text/css"  type="text/javascript" rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.2/css/bootstrap.min.css" />
-  <link type="text/css  type="text/javascript" rel="stylesheet" href="/static/filemanager/bower_components/angular-filemanager/dist/angular-filemanager.css">
+  <link type="text/css"  type="text/javascript" rel="stylesheet" href="/static/filemanager/bower_components/angular-filemanager/dist/angular-filemanager.css">
 </head>
 
 <body>
